+++ lib/functions/copyAssets.mjs
@@ -1,8 +1,5 @@
-#!/usr/bin/env node
-
 import { cp, readdir } from 'fs/promises';
-import { EMOJI } from './constants.mjs';
-import { logError, logMessage } from './loggers.mjs';
+import { EMOJI, logError, logMessage } from '../helpers/index.mjs';
 
 const copyAssets = async (input, output, files) => {
   try {
