+++ lib/functions/index.mjs
@@ -1,0 +1,7 @@
+import copyAssets from './copyAssets.mjs';
+import preflight from './preflight.mjs';
+import processcss from './processcss.mjs';
+import processhtml from './processhtml.mjs';
+import processjs from './processjs.mjs';
+
+export { copyAssets, preflight, processcss, processhtml, processjs };
