+++ lib/functions/preflight.mjs
@@ -1,6 +1,5 @@
 import { mkdir, readdir, rm } from 'fs/promises';
-import { EMOJI } from './constants.mjs';
-import { logError, logMessage } from './loggers.mjs';
+import { EMOJI, logError, logMessage } from '../helpers/index.mjs';
 
 const preflight = async (output) => {
   try {
