+++ lib/functions/processcss.mjs
@@ -4,8 +4,7 @@ import postcss from 'postcss';
 import atImport from 'postcss-import';
 import postcssPresetEnv from 'postcss-preset-env';
 import url from 'postcss-url';
-import { EMOJI, OPTIONS } from './constants.mjs';
-import { logError, logMessage } from './loggers.mjs';
+import { EMOJI, logError, logMessage, OPTIONS } from '../helpers/index.mjs';
 
 const processcss = async (input, output, file) => {
   logMessage(`${EMOJI.artistPalette} Processing styles`);
