+++ lib/functions/processhtml.mjs
@@ -2,8 +2,7 @@ import { readFile, writeFile } from 'fs/promises';
 import htmlnano from 'htmlnano';
 import posthtml from 'posthtml';
 import { hash } from 'posthtml-hash';
-import { EMOJI, OPTIONS } from './constants.mjs';
-import { logError, logMessage } from './loggers.mjs';
+import { EMOJI, logError, logMessage, OPTIONS } from '../helpers/index.mjs';
 
 const processhtml = async (input, output, file) => {
   logMessage(`${EMOJI.fileCabinet} Processing HTML`);
