+++ lib/functions/processjs.mjs
@@ -2,8 +2,7 @@ import { transformFromAstAsync } from '@babel/core';
 import { parse } from '@babel/parser';
 import { readFile, writeFile } from 'fs/promises';
 import { rollup } from 'rollup';
-import { EMOJI, OPTIONS } from './constants.mjs';
-import { logError, logMessage } from './loggers.mjs';
+import { EMOJI, logError, logMessage, OPTIONS } from '../helpers/index.mjs';
 
 const bundlejs = async (input, output, file) => {
   try {
