+++ lib/helpers/loggers.mjs
@@ -1,5 +1,5 @@
 import { exit, stderr, stdout } from 'process';
-import { EMOJI } from './constants.mjs';
+import { EMOJI } from './index.mjs';
 
 export const logError = (error) => {
   stderr.write(`${EMOJI.noEntry} An error has occurred`);
