+++ lib/index.mjs
@@ -4,13 +4,14 @@ import { Command } from 'commander';
 import { readFile } from 'fs/promises';
 import { extname } from 'path';
 import posthtml from 'posthtml';
-import { EMOJI } from './constants.mjs';
-import copyAssets from './copyAssets.mjs';
-import { logError, logMessage } from './loggers.mjs';
-import preflight from './preflight.mjs';
-import processcss from './processcss.mjs';
-import processhtml from './processhtml.mjs';
-import processjs from './processjs.mjs';
+import {
+  copyAssets,
+  preflight,
+  processcss,
+  processhtml,
+  processjs
+} from './functions/index.mjs';
+import { EMOJI, logError, logMessage } from './helpers/index.mjs';
 
 (async () => {
   const program = new Command();
